Create books-choices.rb and add book methods
